plug in error reporting
