+++ ducktape/src/main/scala/io/github/arainko/ducktape/AppliedBuilder.scala
@@ -1,8 +1,8 @@
 package io.github.arainko.ducktape
 
-import io.github.arainko.ducktape.internal.Transformations
+import io.github.arainko.ducktape.internal.TotalTransformations
 
 final class AppliedBuilder[Source, Dest](value: Source) {
   inline def transform(inline config: Field[Source, Dest] | Case[Source, Dest]*): Dest =
-    Transformations.between[Source, Dest](value, "transformation", config*)
+    TotalTransformations.between[Source, Dest](value, "transformation", config*)
 }
