+++ ducktape/src/main/scala/io/github/arainko/ducktape/DefinitionBuilder.scala
@@ -1,8 +1,8 @@
 package io.github.arainko.ducktape
 
-import io.github.arainko.ducktape.internal.Transformations
+import io.github.arainko.ducktape.internal.TotalTransformations
 
 final class DefinitionBuilder[Source, Dest] {
   inline def build(inline config: Field[Source, Dest] | Case[Source, Dest]*): Transformer[Source, Dest] = source =>
-    Transformations.between[Source, Dest](source, "definition", config*)
+    TotalTransformations.between[Source, Dest](source, "definition", config*)
 }
