+++ ducktape/src/main/scala/io/github/arainko/ducktape/Transformer.scala
@@ -1,14 +1,14 @@
 package io.github.arainko.ducktape
 
 import io.github.arainko.ducktape.DefinitionViaBuilder.PartiallyApplied
-import io.github.arainko.ducktape.internal.Transformations
+import io.github.arainko.ducktape.internal.TotalTransformations
 import io.github.arainko.ducktape.internal.FallibleTransformations
 
 trait Transformer[Source, Dest] extends Transformer.Derived[Source, Dest]
 
 object Transformer {
   inline given derive[Source, Dest]: Transformer.Derived[Source, Dest] = new {
-    def transform(value: Source): Dest = Transformations.between[Source, Dest](value, "definition")
+    def transform(value: Source): Dest = TotalTransformations.between[Source, Dest](value, "definition")
   }
 
   def define[Source, Dest]: DefinitionBuilder[Source, Dest] =
